requirement.txt
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
contourpy==1.0.5
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
Django==4.1.7
django-cors-headers==3.11.0
djangorestframework==3.14.0
filelock==3.9.0
fonttools==4.37.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parsel==1.7.0
Pillow==9.2.0
plotly==5.13.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.3.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.1.2
scipy==1.9.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.1.0
threadpoolctl==3.1.0
tldextract==3.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
w3lib==2.1.1
zope.interface==5.5.2
